fixed AddCategory story and update prop to model
