+++ .changeset/cuddly-wombats-rush.md
@@ -1,0 +1,5 @@
+---
+"ui-library": minor
+---
+
+Updated AddCategory story to be responsive
