+++ .changeset/rich-weeks-smile.md
@@ -1,0 +1,5 @@
+---
+"@inkbeard/budget-it": minor
+---
+
+Updated `isActive` local ref to be `isAdding` model
