--- apps/ui-library/src/stories/AddCategories.stories.js
@@ -1,45 +1,0 @@
-import { AddCategory } from '@inkbeard/budget-it';
-
-// More on how to set up stories at: https://storybook.js.org/docs/vue/writing-stories/introduction
-export default {
-  title: 'Budget It/AddCategory',
-  component: AddCategory,
-  tags: ['autodocs'],
-  parameters: {
-    layout: 'padded',
-  },
-};
-export const AddCategoryComponent = {
-  args: {
-    categoryList: [
-      {
-        name: 'Entertainment',
-        id: 1,
-      },
-      {
-        name: 'Food',
-        id: 2,
-      },
-      {
-        name: 'Housing',
-        id: 3,
-      },
-      {
-        name: 'Transportation',
-        id: 4,
-      },
-      {
-        name: 'Utilities',
-        id: 5,
-      },
-      {
-        name: 'Clothing',
-        id: 6,
-      },
-      {
-        name: 'Medical',
-        id: 7,
-      },
-    ],
-  },
-};
